add Account Module

+++ database/migrations/2021_03_15_115045_create_transactions_table.php
@@ -18,7 +18,7 @@ class CreateTransactionsTable extends Migration
             $table->bigInteger('transaction_type_id');
             $table->bigInteger('account_id');
             $table->char('category_id', 5);
-            $table->integer('amount');
+            $table->float('amount', 9, 2);
             $table->char('currency_id', 3);
             $table->dateTime('transaction_date');
             $table->string('transaction_note', 100)->nullable(true);
